messaging_app/requirements.txt
aiosqlite
djangorestframework-simplejwt
